Filled in sections 4.1 and 4.2
Using a temporary document created using an old TeX assignment template
from the University. All other sections other than 4.1 and 4.2 have been
filled in using generic lorem ipsum text. If you wish to use the
document as base of operations, simply remove the command \lipsum from
your section.

Also note that when building, there is an error with the .tex file. I
haven't figured out where the issue is so far, however it is not
impairing the ability to build the pdf file.